moved singleton to single file due to circular import, bind chroma with singleton class
