+++ realtime_ai_character/audio/text_to_speech/unreal_speech.py
@@ -3,7 +3,7 @@ import types
 import httpx
 
 from realtime_ai_character.logger import get_logger
-from realtime_ai_character.utils import Singleton
+from realtime_ai_character.singleton import Singleton
 from realtime_ai_character.audio.text_to_speech.base import TextToSpeech
 
 logger = get_logger(__name__)
