+++ realtime_ai_character/database/chroma.py
@@ -4,6 +4,7 @@ from langchain.vectorstores import Chroma as LangChainChroma
 from langchain.embeddings import OpenAIEmbeddings
 from realtime_ai_character.logger import get_logger
 from .base import Database
+from realtime_ai_character.singleton import Singleton
 
 load_dotenv()
 logger = get_logger(__name__)
@@ -13,13 +14,14 @@ if os.getenv('OPENAI_API_TYPE') == 'azure':
     embedding = OpenAIEmbeddings(openai_api_key=os.getenv("OPENAI_API_KEY"), deployment=os.getenv(
         "OPENAI_API_EMBEDDING_DEPLOYMENT_NAME", "text-embedding-ada-002"), chunk_size=1)
 
-class Chroma(Database):
+class Chroma(Singleton, Database):
     def __init__(self):
         self.db = LangChainChroma(
             collection_name='llm',
             embedding_function=embedding,
             persist_directory='./chroma.db'
         )
+        print("There are", self.db._collection.count(), "in the collection")
     
     def delete_collection(self):
         self.db.delete_collection()
